+++ Sources/StreamChat/ChatClient_Mock.swift
@@ -13,7 +13,8 @@ extension _ChatClient {
             config: .init(apiKey: .init(.unique)),
             workerBuilders: [],
             eventWorkerBuilders: [],
-            environment: .mock
+            environment: .mock,
+            tokenExpirationRetryStrategy: DefaultReconnectionStrategy()
         )
     }
     
@@ -63,7 +64,8 @@ class ChatClientMock<ExtraData: ExtraDataTypes>: _ChatClient<ExtraData> {
         tokenProvider: TokenProvider? = nil,
         workerBuilders: [WorkerBuilder] = [],
         eventWorkerBuilders: [EventWorkerBuilder] = [],
-        environment: Environment = .mock
+        environment: Environment = .mock,
+        tokenExpirationRetryStrategy: WebSocketClientReconnectionStrategy = DefaultReconnectionStrategy()
     ) {
         init_config = config
         init_tokenProvider = tokenProvider
@@ -76,7 +78,8 @@ class ChatClientMock<ExtraData: ExtraDataTypes>: _ChatClient<ExtraData> {
             tokenProvider: tokenProvider,
             workerBuilders: workerBuilders,
             eventWorkerBuilders: eventWorkerBuilders,
-            environment: environment
+            environment: environment,
+            tokenExpirationRetryStrategy: tokenExpirationRetryStrategy
         )
     }