+++ Sources/StreamChat/ChatClient_Tests.swift
@@ -10,6 +10,7 @@ import XCTest
 class ChatClient_Tests: StressTestCase {
     var userId: UserId!
     private var testEnv: TestEnvironment<NoExtraData>!
+    private var time: VirtualTime!
     
     // A helper providing ChatClientConfig with in-memory DB option
     var inMemoryStorageConfig: ChatClientConfig {
@@ -31,6 +32,8 @@ class ChatClient_Tests: StressTestCase {
         super.setUp()
         userId = .unique
         testEnv = .init()
+        time = VirtualTime()
+        VirtualTimeTimer.time = time
     }
     
     override func tearDown() {
@@ -392,7 +395,7 @@ class ChatClient_Tests: StressTestCase {
         XCTAssertNil(providedConnectionId)
     }
     
-    func test_client_webSocketIsDisconnected_becauseTokenExpired_callsReloadUserIfNeeded() throws {
+    func test_webSocketIsDisconnected_becauseTokenExpired_newTokenIsExpiredToo() throws {
         // Create a new chat client
         let client = ChatClient(
             config: inMemoryStorageConfig,
@@ -426,9 +429,29 @@ class ChatClient_Tests: StressTestCase {
                 didUpdateConnectionState: .disconnected(error: error)
             )
         
+        time.run(numberOfSeconds: 0.6)
         // Was called one more time on receiving token expired error
+        AssertAsync.willBeEqual(testEnv.clientUpdater!.reloadUserIfNeeded_callsCount, 2)
+        
+        // Token is expired again
+        testEnv.webSocketClient?
+            .connectionStateDelegate?
+            .webSocketClient(
+                testEnv.webSocketClient!,
+                didUpdateConnectionState: .disconnected(error: error)
+            )
+        
+        // Does not call secondary token refresh right away
         XCTAssertEqual(testEnv.clientUpdater!.reloadUserIfNeeded_callsCount, 2)
         
+        // Does not call secondary token refresh when not enough time has passed
+        time.run(numberOfSeconds: 0.1)
+        XCTAssertEqual(testEnv.clientUpdater!.reloadUserIfNeeded_callsCount, 2)
+        
+        // Calls secondary token refresh when enough time has passed
+        time.run(numberOfSeconds: 3)
+        AssertAsync.willBeEqual(testEnv.clientUpdater!.reloadUserIfNeeded_callsCount, 3)
+
         // We set connectionId to nil after token expiration disconnect
         XCTAssertNil(client.connectionId)
     }
@@ -1147,7 +1170,8 @@ private class TestEnvironment<ExtraData: ExtraDataTypes> {
             backgroundTaskSchedulerBuilder: {
                 self.backgroundTaskScheduler = MockBackgroundTaskScheduler()
                 return self.backgroundTaskScheduler!
-            }
+            },
+            timerType: VirtualTimeTimer.self
         )
     }()
 }