+++ Sources/StreamChatUI/ChatMessageList/ChatMessageListVC_Tests.swift
@@ -13,7 +13,10 @@ final class ChatMessageListVC_Tests: XCTestCase {
     
     override func setUp() {
         super.setUp()
+        var components = Components()
+        components.channelHeaderView = ChatMessageListHeaderView_Mock.self
         vc = ChatMessageListVC()
+        vc.components = components
         channelControllerMock = ChatChannelController_Mock.mock()
         vc.channelController = channelControllerMock
     }
@@ -148,5 +151,19 @@ final class ChatMessageListVC_Tests: XCTestCase {
             isEmbeddedInNavigationController: true,
             variants: [.defaultLight]
         )
+    }
+}
+
+private class ChatMessageListHeaderView_Mock: ChatChannelHeaderView {
+    override var currentUserId: UserId? {
+        .unique
+    }
+    
+    override func setUp() {
+        super.setUp()
+
+        let mockedChannelController = ChatChannelController_Mock<ExtraData>.mock()
+        mockedChannelController.channel_mock = .mock(cid: .unique)
+        channelController = mockedChannelController
     }
 }